修改构建命令丹尼斯未构建

+++ docker/UpdateAll.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 # 部分更新就是保留配置文件的更新
+# shellcheck disable=SC2034
+get_arch='arch'
 # 进入/root目录
 cd /root/ || exit
 # 下载
@@ -8,25 +10,34 @@ wget https://xgzq.tk/library/qlva.tgz
 tar -zxvf qlva.tgz
 # 如果是1保留配置文件
 if test "$1" = "1"; then
-   echo "保留配置文件更新项目"
-   # 把配置文件复制出来
-   cp -f /val/conn.yml /
-   cp -f /val/repeat.sqlite /
-   # 删除ip下所有文件
-   source ./dockergi.sh
-   rm -rf /val/*
-   # 移动文件
+  echo "保留配置文件更新项目"
+  # 把配置文件复制出来
+  cp -f /val/conn.yml /
+  cp -f /val/repeat.sqlite /
+  rm -rf /val/*
+  # 删除ip下所有文件
+  if [[ $get_arch =~ "x86_64" ]];then
+    source ./dockergi.sh
+    # 移动文件
     cp -rf /root/QL_variable/dist/fsbot/* /val
-   # 把配置文件移动到项目目录
-   mv -f /conn.yml /val
-   mv -f /repeat.sqlite /val
+  else
+    cp -rf /root/QL_variable/* /val
+  fi
+  # 把配置文件移动到项目目录
+  mv -f /conn.yml /val
+  mv -f /repeat.sqlite /val
 else
   echo "删除配置文件更新项目"
   # 删除ip下所有文件
   rm -rf /val/*
-  source ./dockergi.sh
-  # 移动文件
-  cp -rf /root/QL_variable/dist/fsbot/* /val
+  if [[ $get_arch =~ "x86_64" ]];then
+    source ./dockergi.sh
+    rm -rf /val/*
+    # 移动文件
+    cp -rf /root/QL_variable/dist/fsbot/* /val
+  else
+     cp -rf /root/QL_variable/* /val
+  fi
 fi
 # 判断文件是否存在存在则执行
 if [ -f "/root/QL_variable/test.sh" ]; then
@@ -38,4 +49,8 @@ fi
 rm -rf /root/qlva.tgz
 rm -rf /root/QL_variable
 # shellcheck disable=SC2046
-kill -9 $(netstat -nlp | grep fsbot | awk '{print $7}' | awk -F"/" '{ print $1 }')
+if [[ $get_arch =~ "x86_64" ]];then
+  kill -9 $(netstat -nlp | grep fsbot | awk '{print $7}' | awk -F"/" '{ print $1 }')
+else
+  kill -9 $(netstat -nlp | grep fsbot.py | awk '{print $7}' | awk -F"/" '{ print $1 }')
+fi
